// src/app/models/Student.ts
export interface Student {
  id: number;
  matricula: string;
  telefono: string;
  fechaNacimiento: string;
  carrera: string;
  semestre: number;
  fechaIngreso: string;
  activo: boolean;
  usuarioId: number;
  firstName: string;
  lastName: string;
  username: string;
  email: string;
  fechaCreacion: string;
  ultimoAcceso: string | null;
  nombreCompleto: string;
  tieneInformacionFinanciera: boolean;
} 